requirements.txt
# Cloud Functions Framework
functions-framework==3.*


# Google Cloud Services
google-cloud-storage==2.14.*
google-cloud-firestore==2.14.*
google-cloud-pubsub==2.19.*
google-cloud-texttospeech==2.14.*
google-cloud-iam==2.*

# Google Generative AI (Gemini)
google-generativeai>=0.8.3

# Flask for local development
Flask==3.0.*
flask-cors==4.*

# Utilities
python-dotenv==1.0.*
pypdf==3.*
requests
